+++ rotkehlchen/chain/ethereum/modules/curve/crvusd/decoder.py
@@ -23,6 +23,7 @@ from rotkehlchen.chain.evm.decoding.structures import (
 )
 from rotkehlchen.globaldb.cache import globaldb_get_general_cache_values
 from rotkehlchen.globaldb.handler import GlobalDBHandler
+from rotkehlchen.history.events.structures.evm_event import EvmProduct
 from rotkehlchen.history.events.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.types import CacheType, ChecksumEvmAddress
@@ -51,6 +52,7 @@ class CurvecrvusdDecoder(CurveBorrowRepayCommonDecoder, ReloadableDecoderMixin):
             evm_inquirer=evm_inquirer,
             base_tools=base_tools,
             msg_aggregator=msg_aggregator,
+            evm_product=EvmProduct.MINTING,
         )
         self.crvusd = EvmToken('eip155:1/erc20:0xf939E0A03FB07F59A73314E73794Be0E57ac1b4E')
 
@@ -172,7 +174,8 @@ class CurvecrvusdDecoder(CurveBorrowRepayCommonDecoder, ReloadableDecoderMixin):
                 event.event_subtype = HistoryEventSubType.DEPOSIT_ASSET
                 event.notes = f'Deposit {event.amount} {collateral_token.symbol} into a leveraged Curve position'  # noqa: E501
                 event.counterparty = CPT_CURVE
-                event.extra_data = {'vault_controller': controller_address}
+                event.product = self.evm_product
+                event.extra_data = {'controller_address': controller_address}
                 break
 
         return DEFAULT_DECODING_OUTPUT